+++ backend/src/main/java/calculator/atoms/AngleMode.java
@@ -1,0 +1,21 @@
+package calculator.atoms;
+
+/**
+ * Enumeration of the representations of angles.
+ *
+ */
+public enum AngleMode {
+
+    /**
+     * The degree representation (180°, 360°, etc.)
+     *
+     */
+    DEG,
+
+    /**
+     * The radian representation (π, π/2, etc.)
+     *
+     */
+    RAD
+
+}